Fix reviewer issues: add assertions to CLI tests, fix code quality
- Add proper assertions to all cli_test.py tests using capsys to verify
  actual output content (MAJOR issue: tests had no assertions)
- Remove dead code DEFAULT_POLL_INTERVAL_SECONDS constant
- Fix _resolve_by_name to reuse _is_agent_match/_is_host_match helpers
  instead of duplicating matching logic
- Move test_wait.py tests into api_test.py (correct naming convention)
- Change polling error log from debug to warning for visibility
